added type methods in Quick Documentation
Methods will only be shown when hovering over a type name in its declaration;
when hovering over a name in a type, such as a parameter, they will not be shown.
This is necessary so as not to make the documentation too large, since the methods
are not always needed when viewing the documentation.

- improved highlighting for expression in Quick Documentation

Fixes #441